Move it to _dtype_abbrs
