+++ torch/fx/graph.py
@@ -19,7 +19,7 @@ from typing import Any, Callable, Literal, NamedTuple, Optional, TYPE_CHECKING
 
 import torch
 import torch.utils._pytree as pytree
-from torch._utils import dtype_abbrs
+from torch.utils._dtype_abbrs import dtype_abbrs
 from torch._C import _fx_map_arg as map_arg, _NodeIter
 
 from . import _pytree as fx_pytree
